Rename dma_multicast_v2 experiments and benchmark
