+++ sw/snitch/tests/multicast_benchmark.c
